refactor: mv storage
